+++ src/snapshot_middleware/project.rs
@@ -301,6 +301,11 @@ fn infer_class_name(name: &str, parent_class: Option<&str>) -> Option<Cow<'stati
         if name == "StarterPlayerScripts" || name == "StarterCharacterScripts" {
             return Some(Cow::Owned(name.to_owned()));
         }
+    } else if parent_class == "Workspace" {
+        // Workspace has a special Terrain class inside it
+        if name == "Terrain" {
+            return Some(Cow::Owned(name.to_owned()));
+        }
     }
 
     None